rag_service/requirements.txt
fastapi[standard]
uvicorn
chromadb
mysql-connector-python
python-dotenv
pydantic
ollama
python-multipart
redis
aioredis